perf(odjsonrt): one bounds test per word in the encoder's scans and stores
The same shape the decoder had: the word loops of appendQuotedStream,
appendStringBodyChecked and appendQuotedV2HTML read their words as
Uint64(src[i:]), two bounds tests each, and the digit stores in the
float formatter wrote through PutUint64(buf[end-8:]), likewise. They
now go through load64 and sub-slices of exact extent, and the byte
loops test their index unsigned. Bounds tests in the encoder's files
106 -> 89. Writing every twitter string once per mode (count 5):
ModeHTML 383 -> 326 us (-14.9%: a byte loop, one test per byte gone),
ModeStream and ModeV2 -2%, ModeV2HTML level; the float benchmarks
within noise.

+++ odjsonrt/encode.go
@@ -95,10 +95,10 @@ func appendQuoted(dst []byte, src []byte, escapeHTML, quoted bool) []byte {
 		// time scan was tried here and measured no faster: the escape set is
 		// wide enough that computing the mask costs about what eight table
 		// lookups do, and the CPU pipelines the lookups well.
-		for i < len(src) && safe[src[i]] {
+		for uint(i) < uint(len(src)) && safe[src[i]] {
 			i++
 		}
-		if i >= len(src) {
+		if uint(i) >= uint(len(src)) {
 			break
 		}
 		if !checked && src[i] >= utf8.RuneSelf {
